refactor: replace float with decimal in some models and update RestaurantFactory

+++ backend/database/migrations/2025_04_02_173529_create_restaurants_table.php
@@ -36,11 +36,11 @@ return new class extends Migration
             $table->string('phone_number')->unique();
             $table->string('email')->unique();
             $table->string('vat_id')->nullable()->unique();
-            $table->float('min_amount')->nullable();
-            $table->float('shipping_cost')->nullable();
+            $table->decimal('min_amount', 8, 2)->nullable();
+            $table->decimal('shipping_cost', 8, 2)->nullable();
             $table->string('logo')->nullable();
             $table->string('cover')->nullable();
-            $table->float('discount')->nullable();
+            $table->decimal('discount', 5, 2)->nullable();
             $table->boolean('is_approved')->default(false);
             $table->timestamps();
         });
